Create DOCKER-RESTART.sh
This is just a fancy way to restart all of your current docker containers. 

If you are lazy and have docker containers that don't start themselves upon boot, this is the script for you. This script takes all of the docker container IDs and restarts the containers/services being hosted.